refactor code without changing function v0.1.1
